// DeCATS-backend/MappedSwap_BackEnd/src/@types/IMiningPool.d.ts
export type StakeRewards = ContractEventLog<{
  userAddr: string;
  stakeAmount: string;
  stakeRewardsAmount: string;
  stakeTime: string;
  nodeID: string;
  stakeHash: string;
  0: string;
  1: string;
  2: string;
  3: string;
  4: string;
  5: string;
}>;